pyproject.toml
[project]
name = "mcp-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.4",
    "mcp[cli]>=1.16.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
